Zipflinger: Add getContent facility (ZipRepo)
Problem: getContent is only available in ZipArchive which is an
expensive object to build. Additionally ZipArchive.getContent
returns a ByteBuffer which stresses heap memory and only allows
for 2 GiB payload.

Solution: Create a new way to getContent via ZipRepo. Make it
return an InputSream which is less heap stressing and allow for
payload > 2 GiB.

Test: Unit Test
Bug: NA
Change-Id: Ia4de2202b3e18bc7f3a5e5c3cbd5707ad9f42169