IRR: Make installerd backward compatible by default.
Problem: The new version of the installer, called installerd, is
backward compatible via a new flag --one_shot. This is a problem
with previous version of AS which don't know of this new needed
flag.

Solution: Instead of enabling backward compatible via a flag, the
preferred behavior is to be backward compatible **by default**.
Enabling the new behavior (daemon) is done via an explicit flag
"-daemon".

Test: Unit Tests
Bug: NA
Change-Id: Ica1556e8074ab90c0cfa652d7f317d825990b663