Define and implement the new VariantSelector API.
this interface is used to select variants that users want to apply
customization on. The interface is loosely based on the previous
iteration of the Variant API and will require a visit to the
API council as soon as possible.

Added unit tests, will change all usage of the old variant API
in a subsequent CL.

Test: Unit tests.
Bug: N/A
Change-Id: I57800dbbef44f976d22b01705719abb8f9046f4e