Add lint CLI flags to override severity
This functionality is already available from Gradle via
the lintOptions block, but not the lint command line.
Adding this (and using the same argument names as in
metalava) to provide an immediate workaround for 168261940
(a fix to make it work properly in all lint.xml scenarios
is coming but needs a bit more work).

Test: Included
Bug: 168261940
Change-Id: Ib65ac95eca6c5c00342f6398bd956514be0f4ca2