Fix new PackagingOptions DSL
Previously, the following syntax in build.gradle files resulted in an
error:

android {
    packagingOptions {
        jniLibs {
            useLegacyPackaging = true
        }
    }
}

It threw an error because (1) AGP wasn't creating the instance of
JniLibsPackagingOptionsImpl via DslServices and (2) there was no
jniLibs(action: Action<JniLibsPackagingOptions>) method on
PackagingOptions.

This change fixes those 2 mistakes for JniLibsPackagingOptions,
DexPackagingOptions, and ResourcesPackagingOptions.

This change also adds test coverage for setting useLegacyPackaging for
jniLibs and dex via the variant API.

Bug: 149770867
Test: Many including NativeSoPackagingOptionsTest and KotlinDslTest
Change-Id: Ic93a32929ff0b6ec40cde64b5a086e87e8c36597