Jetifier: Do not replace invalid android.arch dependency
In older versions, android.arch dependencies have the following group
name:
    android.arch.*:*:1.x.y

Starting with version 2, android.arch has been migrated to AndroidX:
   androidx.*:*:2+

There are cases when the user intends to use the AndroidX version, but
accidentally uses the old group name (e.g.,
`android.arch.persistence.room:compiler:2.1.0`).

In those cases, we can just let the build fail saying that the
dependency is invalid instead of replacing it with an AndroidX version
defined by the Jetifier mappings, which could be lower than the
user-intended version and result in obscure failures.

Bug: 168038088
Test: New test in JetifierTest
Change-Id: I035525545d7bf2df2fc34bfaf9f341d387d0f1b1