move all usage of isDebuggable from variantDslInfo.
VariantDslInfo.isDebuggable() should only be used to initialize the
ComponentCreationConfig subtypes and not used afterwards as it can
be changed during the Variant API execution.

The API moved up to ComponentCreationConfig which is probably higher
than I would have expected (VariantCreationConfig) but some tasks
that are processing resources need this flag so it's there for now.
We can have another pass later to settle where these API belong
in the CreationConfig but so far I want to remove all runtime
dependencies on VariantDslInfo.

Test: existing.
Bug: N/A
Change-Id: Ie6cf51ffca4726b93fb2fa3cd6cadf08a0528f57