Add lint check to detect and flag notification trampolines
This lint check currently detects common trampolines
via BroadcastReceivers; if it can tell that the receiver
launches activities, that is reported as an error, and
otherwise it reports a warning that notifications (except
for notification actions) should not launch receivers.

It flags both BroadcastReceivers and Services, but only
BroadcastReceivers are flagged for actually starting
activities.

Test: Unit tests included
Bug: None
Change-Id: Ia04104cf2c51282671cdc7a1f7692dbec8767e45