IRR: Move Executor out of Workspace
Problem: Having the executor as a set/get in the workspace provides
no gain. It forces functions to pass either a workspace or an
executor, sometimes several layers down which bloats the code.
In cases where a dependency on workspace is not desirable or
possible, it also forces an executor to be passed up to where
it is used.

Solution: All executors to be obtained from a static method.

Test: Unit test
Bug: NA
Change-Id: I5bbc8cad1b22cbfbec187aaaa0a6af98004bde88