New lint integration: CheckDependencies support
When checkDependencies is enabled, the new lint integration
now consumes lint models from project dependencies to allow
those sources to be analyzed.

This is a temporary solution, as eventually lint will
perform analysis in each project variant and export results to be
interpreted and merged in to a report.

For this to work with 'java' or 'java-library' projects, the build
author must now apply the standalone lint plugin 'com.android.lint' to
those projects to export the model in order for lint to find those
sources. The failure mode if that is not done could be made clearer.

Bug: 160392650
Test: Android Gradle Plugin lint integration tests that use checkDependencies run with both integrations.
Change-Id: I188a97fe8d065d5cbb0e5e3ded6ac19ca430100d