Remove file.exists check
iml_to_build should not depend on whether files are on disk or not,
but instead let the build fail if they don't exist.

This is a common scenario if there are references to jar files
in bazel-bin, and the user hasn't built those particular bazel
artifacts.

Rather than having iml_to_build fail in such a case,
we'll generate the correct reference, and let the build
fail if some jars are really missing.

Test: they are disabled, working separately on them.
Bug: N/A
Change-Id: Icfa86ec10b3a25892f7c7d5fb6a8a1e1589eb961